fix(seo): make <html lang> dynamic per locale via route group The root layout hardcoded lang="fr" for all pages. Crawlers saw the wrong lang signal on /en/, /zh/, /ja/. Fix: introduce a (redirect) route group so app/[lang]/layout.tsx becomes the true root layout and renders <html lang={lang}> from the URL segment. The client-side document.documentElement.lang patch is removed as redundant.
