+++ src/app/(redirect)/layout.tsx
@@ -1,0 +1,7 @@
+export default function RedirectLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <html>
+      <body>{children}</body>
+    </html>
+  );
+}
