+++ src/app/(redirect)/page.tsx
